Improve github workflow cache
The cache is too big right now (slow to compress and decompress). This PR
reduces the size of the cache from 2.7G to 400M, its compression time from 2m to
20s, and its decompression time from 45s to 5s. This permits doing jobs in
parallel, which is needed to compensate the fact that the CI itself is longer.
The CI without cache is 20m. The CI with the big cache is 8m. The CI with the
small cache is 15m. But when using a matrix to run the checks in parallel it's
10m because of the unit tests.