Correctly remove UART control flow for Nordic
UART control flow was not correctly removed in #391 for Nordic, because
reception may start in the middle of a frame resulting in corrupted data. By
always keeping the UARTE in reception, we avoid corruption. Data may still be
lost when the buffer is full but that's by design of not using control flow.

Fixes #392