Correctly remove UART control flow for Nordic (#401)
UART control flow was not correctly removed in #391 for Nordic, because
reception may start in the middle of a frame resulting in corrupted
data. By always keeping the UARTE in reception, we avoid corruption.
Data may still be lost when the buffer is full but that's by design of
not using control flow.

This implementation does not provide access to the most recent data and
is only meant for streaming scenario when data is constantly received.
See
https://github.com/google/wasefire/pull/401/commits/1e47e9cf3a6b6bd231ba7240b915773d917390f2
for the commit that removes the logic to access most recent data. It
didn't work because STOPRX doesn't seem to work in the presence of a
short from ENDRX to STARTRX.

Fixes #392