Execution of wasm atomic ops with portable_atomics in the interpreter (#513)
Execution up to the point that `atomic.wast` test passes. Notably that
means notify, wait and fence are nor properly implemented yet, but
everything else is. For the last three some basic checks are implemented
though, like out of bounds and memory alignment.

Also changed the submodule to the feature branch with the fixed tests.

---------

Co-authored-by: Julien Cretin <cretin@google.com>