Fix copy_if_changed with multiple sources
When using `copy_if_changed()` from 2 different sources `a.src` and `b.src` to `x.dst` in an ABA
scenario, it is possible that the modification timestamp of `a.src` is not updated on the last run,
and will thus not be copied. We introduce a copy of the source file next to the destination file to
fix this issue.