Fix platform protocol API documentation (#597)
The current implementation in protocol-usb matches the new
documentation. The rationale for the change is to provide quick feedback
to the host that the device is not ready to handle requests yet. This is
not a big change because this part of the board API is not exposed to
the applet API. It is exclusively used by the scheduler. We might want
to do similar (and bigger) changes to other parts of the API (like UART
or USB serial).