Work around GitHub limitations regarding PR reviews
We are affected by those GitHub issues:
- https://github.com/orgs/community/discussions/6292
- https://github.com/orgs/community/discussions/9636

For the first one, we created @ia0-review. For the second one, we have to make sure that a file has
either exactly @ia0 as code owner, or both @ia0 and @ia0-review. This PR is fixing this second
issue. It is also making sure @ia0 is the code owner of the CODEOWNERS file.