Reduce delta_ip by 1 for "if without else" in validation (#696)
Jumping from "if without else" should be 1 step less than jumping from
"if with else". For "if without else", we don't want to jump over the
"end" because `exit_label()` need to be
[called](https://github.com/zhouwfang/wasefire/blob/de6c7e889865736cb1f11877d5d4e1e9f50729a8/crates/interpreter/src/exec.rs#L788)
at the "end". But for "if with else", we want to jump over the "else"
because otherwise we would be forced to jump to the "end" at "else"
(normally, we would reach "else" after executing the true branch of
"if", and we should not execute the "else" branch).

This is equivalent to
https://github.com/google/wasefire/blob/main/crates/interpreter/src/parser.rs#L566-L569,
which will be deprecated with the side table applied in `exec.rs`.

#46

---------

Co-authored-by: Zhou Fang <33002388+zhou-w-fang@users.noreply.github.com>